chore: refactor + clean up bot's code
