+++ bot/package.json
@@ -6,7 +6,7 @@
     "typecheck": "tsc --noEmit",
     "start": "tsx src/index.ts",
     "dev": "tsx watch --clear-screen=false --include '../server/src/api/routes/private/*' --include ../survev-config.hjson src/index.ts",
-    "register-commands": "tsx src/register-commands.ts"
+    "register-commands": "tsx src/scripts/register-commands.ts"
   },
   "dependencies": {
     "discord.js": "^14.22.1",
