+++ bot/src/components.ts
@@ -55,6 +55,12 @@ export function createSelectUI(matchingPlayers: DropdownPlayer[], searchName: st
     return { embed, row };
 }
 
+
+export const BUTTON_PREFIXES = {
+    BAN_FOR_CHEATING: `search_player_ban_for_cheating_`,
+    BAN_FOR_BAD_NAME: `search_player_ban_for_name_`,
+} as const;
+
 export async function createDiscordDropdownUI(
     interaction: RepliableInteraction,
     matchingPlayers: DropdownPlayer[],
