+++ server/src/api/db/drizzle/meta/_journal.json
@@ -15,6 +15,13 @@
       "when": 1756778270320,
       "tag": "0003_cool_skrulls",
       "breakpoints": true
+    },
+    {
+      "idx": 4,
+      "version": "7",
+      "when": 1760904159005,
+      "tag": "0004_regular_aaron_stack",
+      "breakpoints": true
     }
   ]
 }
